pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "trait2gene"
version = "0.1.0"
description = "End-to-end CLI scaffold for running PoPS-style GWAS gene prioritization workflows."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0-only" }
authors = [{ name = "OpenAI Codex" }]
keywords = ["GWAS", "PoPS", "MAGMA", "gene-prioritization", "bioinformatics"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Bio-Informatics",
  "Operating System :: OS Independent",
]
dependencies = [
  "jinja2>=3.1.4",
  "numpy>=1.26.4",
  "packaging>=24.1",
  "pandas>=2.2.2",
  "platformdirs>=4.2.2",
  "pydantic>=2.8.2",
  "pyyaml>=6.0.2",
  "rich>=13.7.1",
  "scikit-learn>=1.5.1",
  "scipy>=1.13.1",
  "typer>=0.12.3",
]

[project.optional-dependencies]
plot = ["matplotlib>=3.9.1", "plotly>=5.23.0"]
genomics = ["pyranges>=0.1.2", "pysam>=0.22.1"]
dev = ["pytest>=8.3.2", "pytest-cov>=5.0.0", "ruff>=0.6.1", "mypy>=1.11.1"]
docs = [
  "sphinx>=7.4.0",
  "myst-parser>=3.0.0",
  "furo>=2024.8.6",
  "sphinx-copybutton>=0.5.2",
]

[project.urls]
Homepage = "https://github.com/your-org/trait2gene"
Repository = "https://github.com/your-org/trait2gene"
Issues = "https://github.com/your-org/trait2gene/issues"

[project.scripts]
trait2gene = "trait2gene.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/trait2gene"]

[tool.hatch.build]
include = [
  "src/trait2gene/data/**",
  "src/trait2gene/report/templates/**",
]
exclude = ["/tests", "/docs"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = ["src/trait2gene/vendor"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]
